refactor: split worklog on components
